✅ Extract options_example_test.go and slog_example_test.go
